Redesign Option Groups with master-detail
